Assistant checkpoint: Actualizar tipos de datos Ancho y Perfil a decimal
Assistant generated file changes:
- tuco.Clases/Models/Llanta.cs: Cambiar tipos de Ancho y Perfil a decimal en modelo Llanta
- GestionLlantera.Web/Models/DTOs/Inventario/LlantaDTO.cs: Cambiar tipos de Ancho y Perfil a decimal en LlantaDTO
- tuco.Clases/DTOs/Inventario/LlantaTomaDTO.cs: Cambiar tipos de Ancho y Perfil a decimal en LlantaTomaDTO
- GestionLlantera.Web/Services/InventarioService.cs: Agregar método auxiliar para conversión decimal
Agregar método auxiliar para conversión decimal
- API/Controllers/InventarioController.cs: Actualizar endpoint de productos públicos para manejar decimal
- GestionLlantera.Web/wwwroot/js/views/Inventario/editarProducto.js: Actualizar JavaScript para manejar valores decimales

---

User prompt:

Necesito que revises todos los metodos get y endpoint porque al cambiar en la base de datos el tipo de columna ya no esta cargarndo el inventario,

+++ GestionLlantera.Web/Models/DTOs/Inventario/LlantaDTO.cs
@@ -9,10 +9,10 @@ namespace GestionLlantera.Web.Models.DTOs.Inventario
         public int ProductoId { get; set; }
 
         [Display(Name = "Ancho")]
-        public int? Ancho { get; set; }
+        public decimal? Ancho { get; set; }
 
         [Display(Name = "Perfil")]
-        public int? Perfil { get; set; }
+        public decimal? Perfil { get; set; }
 
         [Display(Name = "Diámetro")]
         public string? Diametro { get; set; }
